pyproject.toml
[tool.poetry]
name = "raspi-gpio"
version = "0.1.0"
description = "RPi.GPIO wrapper with mocks for developmennt on any platform"
authors = ["Mark Parker <mark@parker-programs.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "raspi_gpio"}]
homepage = "https://pypi.org/project/raspi-gpio/"
repository = "https://github.com/MarkParker5/raspi-gpio"
documentation = "https://github.com/MarkParker5/raspi-gpio/blob/main/README.md"
keywords = ["raspberry-pi", "gpio", "rpi", "rpi-gpio", "RPi.GPIO"]

[tool.poetry.dependencies]
python = "^3.10"
rpi-gpio = {version = "^0.7.1", platform = "linux"}
spidev = {version = "^3.6", platform = "linux"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
